feat: started mcp token dialog option
