+++ components.d.ts
@@ -30,6 +30,9 @@ declare module 'vue' {
     VoIconBin: typeof import('./src/components/icon/VoIconBin.vue')['default']
     VoIconIssues: typeof import('./src/components/icon/VoIconIssues.vue')['default']
     VoListItem: typeof import('./src/components/app/VoListItem.vue')['default']
+    VoMcpDialog: typeof import('./src/components/mcp/VoMcpDialog.vue')['default']
+    VoMcpListItem: typeof import('./src/components/mcp/VoMcpListItem.vue')['default']
+    VoMcpTokenTable: typeof import('./src/components/mcp/VoMcpTokenTable.vue')['default']
     VoNotificationsBadge: typeof import('./src/components/notifications/VoNotificationsBadge.vue')['default']
     VoNotificationsBanner: typeof import('./src/components/notifications/VoNotificationsBanner.vue')['default']
     VoNotificationsBannerList: typeof import('./src/components/notifications/VoNotificationsBannerList.vue')['default']
