+++ src/components/mcp/VoMcpListItem.vue
@@ -1,0 +1,18 @@
+<template>
+  <VoListItem
+    :active="dialog"
+    link
+    :prepend-icon="`svg:${mdiKeyVariant}`"
+    title="MCP Access"
+  >
+    <VoMcpDialog v-model="dialog" />
+  </VoListItem>
+</template>
+
+<script lang="ts" setup>
+
+// Icons
+  import { mdiKeyVariant } from '@mdi/js'
+
+  const dialog = shallowRef(false)
+</script>
